requirements.txt
langchain-core==0.2.4
langgraph==0.0.64
langchain-community==0.2.3
langchain-openai==0.1.8
termcolor==2.4.0
chainlit==1.1.202
psycopg 
psycopg-pool 
langgraph-checkpoint-postgres
fastapi 
uvicorn
psycopg2-binary
